+++ src/Html2Pdf.php
@@ -232,7 +232,7 @@ class Html2Pdf
         return array(
             'major'     => 5,
             'minor'     => 0,
-            'revision'  => 2,
+            'revision'  => 3,
         );
     }
 
@@ -3144,9 +3144,36 @@ class Html2Pdf
      */
     protected function _tag_open_DIV($param, $other = 'div')
     {
+        if ($this->parsingCss->value['page-break-before'] == "always") {
+              $this->_setNewPage(null, '', null, $this->_defaultTop);
+              $this->parsingCss->setPosition();
+        }
+
+        if ($this->parsingCss->value['page-break-inside'] == "avoid") {
+            $this->_maxH = 0;
+
+            // create a sub HTML2PDF to execute the content of the tag, to get the dimensions
+            $sub = $this->createSubHTML();
+            $sub->parsingHtml->code = $this->parsingHtml->getLevel($this->_parsePos);
+            $sub->_makeHTMLcode();
+            $y = $this->pdf->getY();
+
+            // if the content does not fit on the page => new page
+            if (
+                $sub->_maxY < ($this->pdf->getH() - $this->pdf->gettMargin()-$this->pdf->getbMargin()) &&
+                $y + $sub->_maxY>=($this->pdf->getH() - $this->pdf->getbMargin())
+            ) {
+                $this->_setNewPage(null, '', null, $this->_defaultTop);
+            }
+
+            // destroy the sub HTML2PDF
+            $this->_destroySubHTML($sub);
+        }
+
         if ($this->_isForOneLine) {
             return false;
         }
+
         if (!is_null($this->debug)) {
             $this->debug->addStep(strtoupper($other), true);
         }
@@ -3553,6 +3580,12 @@ class Html2Pdf
         if ($block) {
             $this->_tag_open_BR(array());
         }
+
+        if ($this->parsingCss->value['page-break-after'] == "always") {
+              $this->_setNewPage(null, '', null, $this->_defaultTop);
+              $this->parsingCss->setPosition();
+        }
+
         if (!is_null($this->debug)) {
             $this->debug->addStep(strtoupper($other), false);
         }