+++ src/Parsing/Css.php
@@ -181,6 +181,10 @@ class Css
 
         $this->value['xc'] = null;
         $this->value['yc'] = null;
+
+        $this->value['page-break-before'] = null;
+        $this->value['page-break-inside'] = null;
+        $this->value['page-break-after']  = null;
     }
 
     /**
@@ -1185,7 +1189,11 @@ class Css
                 case 'start':
                     $this->value[$nom] = intval($val);
                     break;
-
+                case 'page-break-before':
+                case 'page-break-inside':
+                case 'page-break-after':
+                    $this->value[$nom] = $val;
+                    break;
                 default:
                     break;
             }